Move React DND stuff to appropriate folders
